Backend/requirements.txt
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
firebase_admin==6.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gdown
networkx
sentencepiece
fastapi[uvicorn]
# PyPDF2
matplotlib
uvicorn
cloud-sql-python-connector